extract `extractCode`
