+++ src/index.js
@@ -1,13 +1,5 @@
 const fetchMessages = require('./fetchmessages');
-const numberRegex = /([\d{4}]+)/;
-
-const extractCode = (message) => {
-  const code = message.files[0].plain_text.match(numberRegex);
-
-  if (code != null) {
-    return code[0];
-  }
-};
+const extractCode = require('./extractcode');
 
 exports.getShedCode = async (req, res) => {
   const message = await fetchMessages();
